SeeQR: Version 2

+++ backend/main.ts
@@ -1,13 +1,14 @@
 // Import parts of electron to use
 import { app, BrowserWindow, ipcMain, Menu } from 'electron';
+import { appendFile } from 'fs/promises';
 import { join } from 'path';
 import { format } from 'url';
-import './channels' // all channels live here
+//import './channels' // all channels live here
+import execute from './channels';
 
 const { exec } = require('child_process');
 const appMenu = require('./mainMenu'); // use appMenu to add options in top menu bar of app
 const path = require('path');
-const createInsertQuery = require('./dummy_db/dataGenHandler');
 
 /************************************************************
  *********** PACKAGE ELECTRON APP FOR DEPLOYMENT ***********
@@ -23,6 +24,9 @@ const createInsertQuery = require('./dummy_db/dataGenHandler');
 // Keep a global reference of the window objects, if you don't, the window will
 // be closed automatically when the JavaScript object is garbage collected.
 let mainWindow: any;
+
+//global variable to determine whether or not containers are still running
+let pruned: boolean = false;
 
 let mainMenu = Menu.buildFromTemplate(require('./mainMenu'));
 // Keep a reference for dev mode
@@ -73,6 +77,9 @@ function createWindow() {
   // Don't show until we are ready and loaded
   mainWindow.once('ready-to-show', (event) => {
     mainWindow.show();
+    // uncomment code below before running production build and packaging
+    // const yamlPath = join(__dirname, '../../docker-compose.yml')
+    // const runDocker: string = `docker-compose -f '${yamlPath}' up -d`;
     const runDocker: string = `docker-compose up -d`;
     exec(runDocker, (error, stdout, stderr) => {
       if (error) {
@@ -86,28 +93,33 @@ function createWindow() {
       console.log(`${stdout}`);
     })
   });
-
-  // Emitted when the window is closed.
-  mainWindow.on('closed', function () {
-    // Stop and remove postgres-1 and busybox-1 Docker containers upon window exit.
-    const pruneContainers: string = 'docker rm -f postgres-1 busybox-1';
-    const executeQuery = (str) => {
-      exec(str, (error, stdout, stderr) => {
-        if (error) {
-          console.log(`error: ${error.message}`);
-          return;
-        }
-        if (stderr) {
-          console.log(`stderr: ${stderr}`);
-          return;
-        }
-        console.log(`${stdout}`);
-      })
-    };
-    executeQuery(pruneContainers);
-    mainWindow = null;
-  });
 }
+
+app.on('before-quit', (event: any) => {
+  // check if containers have already been pruned--else, continue with default behavior to terminate application
+  if (!pruned) {
+    event.preventDefault();
+    // Stop and remove postgres-1 and busybox-1 Docker containers upon window exit.
+    const stopContainers: string = 'docker stop postgres-1 busybox-1';
+    const pruneContainers: string = 'docker rm -f postgres-1 busybox-1';
+    // this command removes the volume which stores the session data for the postgres instance
+    // comment this out for dev
+    const pruneVolumes: string = 'docker volume rm -f seeqr_database-data';
+
+    // use this string for production build
+    // const pruneVolumes: string = 'docker volume rm -f app_database-data'
+
+    const step4 = () => {
+      pruned = true;
+      app.quit()
+    };
+    const step3 = () => execute(pruneVolumes, step4);
+    const step2 = () => execute(pruneContainers, step3);
+
+    execute(stopContainers, step2);
+  }
+})
+
 
 // Invoke createWindow to create browser windows after Electron has been initialized.
 // Some APIs can only be used after this event occurs.
@@ -130,3 +142,4 @@ app.on('activate', () => {
   }
 });
 
+export default mainWindow;
